data_processing/requirements.txt
# Core data handling and manipulation
pandas>=1.5.0,<2.3.0
numpy>=1.21.0,<1.27.0 # Often a dependency of pandas/sentence-transformers, good to specify

# Downloading data from URL
requests>=2.28.0,<2.33.0

# Google Cloud Storage interaction
google-cloud-storage>=2.5.0,<2.18.0

# MongoDB interaction (using the standard synchronous driver for simplicity in scripts)
pymongo>=4.0.0,<5.0.0

# Embeddings generation using Hugging Face models
sentence-transformers>=2.2.0,<3.0.0
# Note: sentence-transformers will install a backend like torch or tensorflow
# Specify one if needed, e.g., torch>=1.9.0,<2.4.0
# If running on CPU-only machine, installation might be smaller.

# Environment variable loading (for local execution using .env file)
python-dotenv>=0.20.0,<1.1.0

# Optional: Redis client (if update_recommendations.py writes to Redis)
# redis>=4.3.0,<5.1.0

# Optional: Progress bars for long tasks like embedding generation
# tqdm>=4.60.0,<5.0.0